updated all files before implementing ui on 4th of may 2024

+++ app/src/main/java/com/sagar/nourishnow/domain/model/Nutrient.kt
@@ -8,7 +8,7 @@ data class Nutrient(
     val  name: String,
     val unit: MeasuringUnit
 ){
-    fun toNutrientOffline(majorNutrientId: Int): NutrientOffline {
+    fun toNutrientOffline(majorNutrientId: Long): NutrientOffline {
         return NutrientOffline(
             quantity = quantity,
             name = name,
